V2.8

+++ pages/manageFormula.php
@@ -160,6 +160,12 @@ if($_POST['action'] == 'addFormula'){
 //DELETE FORMULA
 if($_GET['action'] == 'delete' && $_GET['fid']){
 	$fid = mysqli_real_escape_string($conn, $_GET['fid']);
+	
+	if(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM formulasMetaData WHERE fid = '$fid' AND isProtected = '1'"))){
+		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error:</strong> formula '.base64_decode($fid).' is protected.</div>';
+		return;
+	}
+	
 	if(mysqli_query($conn, "DELETE FROM formulas WHERE fid = '$fid'")){
 		mysqli_query($conn, "DELETE FROM formulasMetaData WHERE fid = '$fid'");
 		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>Formula '.base64_decode($fid).' deleted!</div>';
@@ -259,9 +265,11 @@ if($_GET['action'] == 'removeFromCart' && $_GET['materialId']){
 
 //PRINTING
 if($_GET['action'] == 'printLabel' && $_GET['name']){
+	if (file_exists(__ROOT__.'/tmp/labels/') === FALSE) {
+		mkdir(__ROOT__.'/tmp/labels/', 0740, true);
+	}
 	$name = $_GET['name'];
 		
-
 	if($settings['label_printer_size'] == '62' || $settings['label_printer_size'] == '62 --red'){
 		
 		if($_GET['batchID']){
@@ -287,7 +295,7 @@ if($_GET['action'] == 'printLabel' && $_GET['name']){
 	imagefilledrectangle($lbl, 0, 0, $w, $h, $white);
 	
 	$text = trim($name.$extras);
-	$font = '../fonts/Arial.ttf';
+	$font = __ROOT__.'/fonts/Arial.ttf';
 
 	imagettftext($lbl, $settings['label_printer_font_size'], 0, 0, 150, $black, $font, $text);
 	$lblF = imagerotate($lbl, 90 ,0);
@@ -300,7 +308,7 @@ if($_GET['action'] == 'printLabel' && $_GET['name']){
 		$extras = ' @'.$_GET['dilution'].'% in '.$_GET['dilutant'];
 		imagettftext($lblF, 40, 90, 200, 600, $black, $font, $extras);
 	}
-	$save = "../tmp/labels/".base64_encode($text.'png');
+	$save = __ROOT__.'/tmp/labels/'.base64_encode($text.'png');
 
 	if(imagepng($lblF, $save)){
 		imagedestroy($lblF);
@@ -314,6 +322,10 @@ if($_GET['action'] == 'printLabel' && $_GET['name']){
 
 //PRINT BOX LABEL
 if($_GET['action'] == 'printBoxLabel' && $_GET['name']){
+	if (file_exists(__ROOT__.'/tmp/labels/') === FALSE) {
+		mkdir(__ROOT__.'/tmp/labels/', 0740, true);
+	}
+	
 	if(empty($_GET['copies']) || !is_numeric($_GET['copies'])){
 		$copies = '1';
 	}else{
@@ -365,7 +377,7 @@ if($_GET['action'] == 'printBoxLabel' && $_GET['name']){
 	imagefilledrectangle($lbl, 0, 0, $h, $w, $white);
 	
 	$text = strtoupper($q['product_name']);
-	$font = '../fonts/Arial.ttf';
+	$font = __ROOT__.'/fonts/Arial.ttf';
 				//font size 15 rotate 0 center 360 top 50
 	imagettftext($lbl, 30, 0, 250, 50, $black, $font, $text);
 	imagettftext($lbl, 25, 0, 300, 100, $black, $font, 'INGREDIENTS');
@@ -374,12 +386,15 @@ if($_GET['action'] == 'printBoxLabel' && $_GET['name']){
 	imagettftext($lblF, 20, 0, 50, 150, $black, $font, wordwrap ($allergenFinal, 60));
 	imagettftext($lblF, 20, 0, 150, 490, $black, $font, wordwrap ($info, 50));
 
-	$save = "../tmp/labels/".base64_encode($text.'png');
+	$save = __ROOT__.'/tmp/labels/'.base64_encode($text.'png');
 
 	if(imagepng($lblF, $save)){
 		imagedestroy($lblF);
+		if($_GET['download'] == '1'){
+			echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><a href="'.'/tmp/labels/'.base64_encode($text.'png').'" target="_blank">Get Label here</a></div>';
+			return;
+		}
 		for ($k = 0; $k < $copies; $k++){
-			//echo '<img src="'.$save.'"/>';
 			shell_exec('/usr/bin/brother_ql -m '.$settings['label_printer_model'].' -p tcp://'.$settings['label_printer_addr'].' print -l '.$settings['label_printer_size'].' '. $save);
 		}
 		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>Print sent!</div>';
@@ -387,71 +402,4 @@ if($_GET['action'] == 'printBoxLabel' && $_GET['name']){
 	return;
 }
 
-//DOWNLOAD BOX LABEL
-if($_GET['action'] == 'downloadBoxLabel' && $_GET['name']){
-		
-		$name = mysqli_real_escape_string($conn, $_GET['name']);
-		$q = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM formulasMetaData WHERE name = '$name'"));
-		$qIng = mysqli_query($conn, "SELECT ingredient FROM formulas WHERE name = '$name'");
-		
-		while($ing = mysqli_fetch_array($qIng)){
-				$chName = mysqli_fetch_array(mysqli_query($conn, "SELECT chemical_name FROM ingredients WHERE name = '".$ing['ingredient']."' AND allergen = '1'"));
-				if($chName['chemical_name']){
-					$getAllergen['name'] = $chName['chemical_name'];
-				}else{
-					$getIngAlergen = mysqli_fetch_array(mysqli_query($conn, "SELECT name FROM ingredients WHERE name = '".$ing['ingredient']."' AND allergen = '1'"));
-					$qAll = mysqli_query($conn, "SELECT name FROM allergens WHERE ing = '".$ing['ingredient']."'");
-					while($getAllergen = mysqli_fetch_array($qAll)){
-						$allergen[] = $getAllergen['name'];
-					}
-				}
-			$allergen[] = $getIngAlergen['name'];
-			$allergen[] = $getAllergen['name'];
-		}
-		$allergen[] = "Denatureted Ethyl Alcohol ".$_GET['carrier']."% Vol, \nFragrance, DPG, Distilled Water";
-
-		if($_GET['batchID']){
-			$bNo = $_GET['batchID'];
-		}else{
-			$bNO = 'N/A';
-		}
-		$allergenFinal = implode(", ",array_filter($allergen));
-		if($settings['brandName']){
-			$brand = $settings['brandName'];
-		}else{
-			$brand = 'PV Pro';
-		}
-		$info = "FOR EXTERNAL USE ONLY. \nKEEP AWAY FROM HEAT AND FLAME. \nKEEP OUT OF REACH OF CHILDREN. \nAVOID SPRAYING IN EYES. \n \nProduction: ".date("d/m/Y")." \nB. NO: ".$bNo." \n$brand";
-		$w = '720';
-		$h = '860';
-	
-		
-	$lbl = imagecreatetruecolor($h, $w);
-
-	$white = imagecolorallocate($lbl, 255, 255, 255);
-	$black = imagecolorallocate($lbl, 0, 0, 0);	
-	
-	imagefilledrectangle($lbl, 0, 0, $h, $w, $white);
-	
-	$text = strtoupper($q['product_name']);
-	$font = '../fonts/Arial.ttf';
-				//font size 15 rotate 0 center 360 top 50
-	imagettftext($lbl, 30, 0, 250, 50, $black, $font, $text);
-	imagettftext($lbl, 25, 0, 300, 100, $black, $font, 'INGREDIENTS');
-	$lblF = imagerotate($lbl, 0 ,0);
-	
-	imagettftext($lblF, 20, 0, 50, 150, $black, $font, wordwrap ($allergenFinal, 60));
-	imagettftext($lblF, 20, 0, 150, 490, $black, $font, wordwrap ($info, 50));
-
-	$save = "../tmp/labels/".base64_encode($text.'png');
-			
-			//echo '<img src="'.$save.'"/>';
-
-	if(imagepng($lblF, $save)){
-		imagedestroy($lblF);
-		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><a href="'.$save.'" target="_blank">Get Label here</a></div>';
-	}
-//}
-	return;
-}
 ?>
